Added pdfkit and report generator classes
