+++ package.json
@@ -29,8 +29,8 @@
     "adm-zip": "^0.5.6",
     "ansi-regex": "^6.0.1",
     "approx-string-match": "^1.1.0",
-    "aws-sdk": "^2.1167.0",
     "async": "^3.2.4",
+    "aws-sdk": "^2.1167.0",
     "axios": "^0.27.2",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
@@ -52,6 +52,7 @@
     "node-cron": "^3.0.0",
     "node-forge": "^1.3.0",
     "node-xlsx": "^0.21.0",
+    "pdfkit": "^0.13.0",
     "pg": "^8.6.0",
     "pg-hstore": "^2.3.3",
     "reflect-metadata": "^0.1.13",
